+++ web-admin/src/features/organizations/users/OrgUsersTable.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-  import { writable } from "svelte/store";
   import type {
     RpcStatus,
     V1ListOrganizationInvitesResponse,
@@ -10,25 +9,13 @@
   import OrgUsersTableUserCompositeCell from "./OrgUsersTableUserCompositeCell.svelte";
   import OrgUsersTableActionsCell from "./OrgUsersTableActionsCell.svelte";
   import OrgUsersTableRoleCell from "./OrgUsersTableRoleCell.svelte";
-  import {
-    createSvelteTable,
-    flexRender,
-    getCoreRowModel,
-    getSortedRowModel,
-  } from "@tanstack/svelte-table";
-  import type {
-    ColumnDef,
-    OnChangeFn,
-    SortingState,
-    TableOptions,
-  } from "@tanstack/svelte-table";
-  import { createVirtualizer } from "@tanstack/svelte-virtual";
-  import ArrowDown from "@rilldata/web-common/components/icons/ArrowDown.svelte";
+  import { flexRender, type ColumnDef } from "@tanstack/svelte-table";
   import type {
     InfiniteData,
     InfiniteQueryObserverResult,
   } from "@tanstack/svelte-query";
   import { ExternalLinkIcon } from "lucide-svelte";
+  import InfiniteScrollTable from "@rilldata/web-admin/components/InfiniteScrollTable.svelte";
 
   interface OrgUser extends V1OrganizationMemberUser, V1OrganizationInvite {
     invitedBy?: string;
@@ -50,21 +37,7 @@
   export let onAttemptRemoveBillingContactUser: () => void;
   export let onAttemptChangeBillingContactUserRole: () => void;
 
-  const ROW_HEIGHT = 69;
-  const OVERSCAN = 5;
-
-  let virtualListEl: HTMLDivElement;
-  let sorting: SortingState = [];
-
   $: safeData = Array.isArray(data) ? data : [];
-  $: {
-    if (safeData) {
-      options.update((old) => ({
-        ...old,
-        data: safeData,
-      }));
-    }
-  }
 
   $: columns = <ColumnDef<OrgUser, any>[]>[
     {
@@ -120,214 +93,33 @@
     },
   ];
 
-  const setSorting: OnChangeFn<SortingState> = (updater) => {
-    if (updater instanceof Function) {
-      sorting = updater(sorting);
-    } else {
-      sorting = updater;
+  function handleLoadMore() {
+    if (usersQuery.hasNextPage) {
+      usersQuery.fetchNextPage();
     }
-
-    options.update((old) => ({
-      ...old,
-      state: {
-        ...old.state,
-        sorting,
-      },
-    }));
-  };
-
-  $: options = writable<TableOptions<OrgUser>>({
-    data: safeData,
-    columns,
-    state: {
-      sorting,
-    },
-    onSortingChange: setSorting,
-    getCoreRowModel: getCoreRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-  });
-
-  $: table = createSvelteTable(options);
-
-  $: rows = $table.getRowModel().rows;
-
-  $: virtualizer = createVirtualizer<HTMLDivElement, HTMLDivElement>({
-    count: 0,
-    getScrollElement: () => virtualListEl,
-    estimateSize: () => ROW_HEIGHT,
-    overscan: OVERSCAN,
-  });
-
-  $: {
-    const hasNextPage = usersQuery.hasNextPage || invitesQuery.hasNextPage;
-
-    $virtualizer.setOptions({
-      count: hasNextPage ? safeData.length + 1 : safeData.length,
-    });
-
-    const [lastItem] = [...$virtualizer.getVirtualItems()].reverse();
-
-    if (
-      lastItem &&
-      lastItem.index > safeData.length - 1 &&
-      hasNextPage &&
-      !usersQuery.isFetchingNextPage &&
-      !invitesQuery.isFetchingNextPage
-    ) {
-      if (usersQuery.hasNextPage) {
-        usersQuery.fetchNextPage();
-      }
-      if (invitesQuery.hasNextPage) {
-        invitesQuery.fetchNextPage();
-      }
+    if (invitesQuery.hasNextPage) {
+      invitesQuery.fetchNextPage();
     }
   }
 
   $: dynamicTableMaxHeight = data.length > 12 ? `calc(100dvh - 300px)` : "auto";
+
+  const headerIcons = {
+    roleName: {
+      icon: ExternalLinkIcon,
+      href: "https://docs.rilldata.com/manage/roles-permissions#organization-level-permissions",
+    },
+  };
 </script>
 
-<!-- FIXME: hoist this to a InfiniteScrollTable component -->
-<div
-  class={`list scroll-container ${dynamicTableMaxHeight}`}
-  bind:this={virtualListEl}
-  style:max-height={dynamicTableMaxHeight}
->
-  <div class="table-wrapper" style="position: relative;">
-    <table>
-      <thead>
-        {#each $table.getHeaderGroups() as headerGroup}
-          <tr class="h-10">
-            {#each headerGroup.headers as header (header.id)}
-              {@const widthPercent = header.column.columnDef.meta?.widthPercent}
-              {@const marginLeft = header.column.columnDef.meta?.marginLeft}
-              <th
-                colSpan={header.colSpan}
-                style={`width: ${widthPercent}%;`}
-                class="px-4 py-2 text-left"
-                on:click={header.column.getToggleSortingHandler()}
-              >
-                {#if !header.isPlaceholder}
-                  <div
-                    style={`margin-left: ${marginLeft};`}
-                    class:cursor-pointer={header.column.getCanSort()}
-                    class:select-none={header.column.getCanSort()}
-                    class="font-semibold text-gray-500 flex flex-row items-center gap-x-1 text-sm"
-                  >
-                    <svelte:component
-                      this={flexRender(
-                        header.column.columnDef.header,
-                        header.getContext(),
-                      )}
-                    />
-                    {#if header.column.id === "roleName"}
-                      <a
-                        href="https://docs.rilldata.com/manage/roles-permissions#organization-level-permissions"
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        class="hover:text-gray-700"
-                      >
-                        <ExternalLinkIcon
-                          class="text-gray-500"
-                          size="11px"
-                          strokeWidth={2}
-                        />
-                      </a>
-                    {/if}
-                    {#if header.column.getIsSorted().toString() === "asc"}
-                      <span>
-                        <ArrowDown flip size="12px" />
-                      </span>
-                    {:else if header.column.getIsSorted().toString() === "desc"}
-                      <span>
-                        <ArrowDown size="12px" />
-                      </span>
-                    {/if}
-                  </div>
-                {/if}
-              </th>
-            {/each}
-          </tr>
-        {/each}
-      </thead>
-      <tbody>
-        {#if $table.getRowModel().rows.length === 0}
-          <tr>
-            <td
-              colspan={columns.length}
-              class="px-4 py-4 text-center text-gray-500"
-            >
-              No users found
-            </td>
-          </tr>
-        {:else}
-          {#each $virtualizer.getVirtualItems() as virtualRow, idx (virtualRow.index)}
-            <tr
-              style="height: {virtualRow.size}px; transform: translateY({virtualRow.start -
-                idx * virtualRow.size}px);"
-            >
-              {#each rows[virtualRow.index]?.getVisibleCells() ?? [] as cell (cell.id)}
-                <td
-                  class={`px-4 py-2 max-w-[200px] truncate ${cell.column.id === "actions" ? "w-1" : ""}`}
-                  data-label={cell.column.columnDef.header}
-                >
-                  <svelte:component
-                    this={flexRender(
-                      cell.column.columnDef.cell,
-                      cell.getContext(),
-                    )}
-                  />
-                </td>
-              {/each}
-            </tr>
-          {/each}
-        {/if}
-      </tbody>
-    </table>
-  </div>
-</div>
-
-<style lang="postcss">
-  table {
-    @apply border-separate border-spacing-0 w-full;
-  }
-  table th,
-  table td {
-    @apply border-b border-gray-200;
-  }
-  thead {
-    @apply sticky top-0 z-30 bg-white;
-  }
-  thead tr th {
-    @apply border-t border-gray-200;
-  }
-  thead tr th:first-child {
-    @apply border-l;
-    @apply rounded-tl-sm;
-  }
-  thead tr th:last-child {
-    @apply border-r;
-    @apply rounded-tr-sm;
-  }
-  thead tr:last-child th {
-    @apply border-b;
-  }
-  tbody tr:first-child {
-    @apply border-t-0;
-  }
-  tbody td:first-child {
-    @apply border-l;
-  }
-  tbody td:last-child {
-    @apply border-r;
-  }
-  tbody tr:last-child td:first-child {
-    @apply rounded-bl-sm;
-  }
-  tbody tr:last-child td:last-child {
-    @apply rounded-br-sm;
-  }
-  .scroll-container {
-    width: 100%;
-    overflow-y: auto;
-  }
-</style>
+<InfiniteScrollTable
+  {data}
+  {columns}
+  hasNextPage={usersQuery.hasNextPage || invitesQuery.hasNextPage}
+  isFetchingNextPage={usersQuery.isFetchingNextPage ||
+    invitesQuery.isFetchingNextPage}
+  onLoadMore={handleLoadMore}
+  maxHeight={dynamicTableMaxHeight}
+  emptyStateMessage="No users found"
+  {headerIcons}
+/>
